+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ✨ Added support for logical operators `&&` and `||`
 - 🐛 Fixed float convertion to string where 0 float was converted to 0.0. Now, float convertion to string works perfectly
 - 🐛 Fixed wrong convertion of ternary expression to string
+- 🐛 Fixed infinite loop bug that whould appear when you have an illegal token like `{{ 4 ~ 3 }}`
 
 ## v2.6.8 (2025-06-19)
 - ✨ Added Container Engine support to the project